+++ src/app/(Dashboard)/Landlord/Components/Grids/types.ts
@@ -8,3 +8,13 @@ export type Complex = {
     createdAt: string;
     modifiedAt: string;
 }
+
+export type Apartment = {
+    id: number;
+    complexId: number;
+    name: string;
+    address: string;
+    description: string;
+    createdAt: string;
+    modifiedAt: string;
+}
